Organized useTabs in its own directory
