+++ src/hooks/useTabs/index.js
@@ -1,0 +1,1 @@
+export { default } from './useTabs';
